+++ R/bayesplot-package.R
@@ -4,8 +4,8 @@
 #' @aliases bayesplot
 #'
 #' @import ggplot2 stats rlang
-#' @importFrom dplyr %>% summarise group_by select
 #' @importFrom lifecycle deprecated deprecate_warn is_present
+#' @importFrom dplyr %>% summarise group_by select
 #'
 #' @description
 #' \if{html}{